Closes #17143: Deprecate the use of dedicated nested API serializers
