+++ netbox/circuits/api/serializers.py
@@ -1,3 +1,2 @@
 from .serializers_.providers import *
 from .serializers_.circuits import *
-from .nested_serializers import *
