+++ netbox/core/api/serializers.py
@@ -1,4 +1,3 @@
 from .serializers_.change_logging import *
 from .serializers_.data import *
 from .serializers_.jobs import *
-from .nested_serializers import *
