+++ netbox/dcim/api/serializers.py
@@ -11,4 +11,3 @@ from .serializers_.devices import *
 from .serializers_.device_components import *
 from .serializers_.power import *
 from .serializers_.rackunits import *
-from .nested_serializers import *
