+++ netbox/extras/api/nested_serializers.py
@@ -1,3 +1,5 @@
+import warnings
+
 from rest_framework import serializers
 
 from extras import models
@@ -19,6 +21,12 @@ __all__ = [
     'NestedTagSerializer',  # Defined in netbox.api.serializers
     'NestedWebhookSerializer',
 ]
+
+# TODO: Remove in v4.2
+warnings.warn(
+    f"Dedicated nested serializers will be removed in NetBox v4.2. Use Serializer(nested=True) instead.",
+    DeprecationWarning
+)
 
 
 class NestedEventRuleSerializer(WritableNestedSerializer):
