+++ netbox/tenancy/api/nested_serializers.py
@@ -1,7 +1,7 @@
-from drf_spectacular.utils import extend_schema_serializer
-from rest_framework import serializers
+import warnings
 
 from netbox.api.serializers import WritableNestedSerializer
+from serializers_.nested import NestedContactGroupSerializer, NestedTenantGroupSerializer
 from tenancy.models import *
 
 __all__ = [
@@ -13,22 +13,16 @@ __all__ = [
     'NestedTenantSerializer',
 ]
 
+# TODO: Remove in v4.2
+warnings.warn(
+    f"Dedicated nested serializers will be removed in NetBox v4.2. Use Serializer(nested=True) instead.",
+    DeprecationWarning
+)
+
 
 #
 # Tenants
 #
-
-@extend_schema_serializer(
-    exclude_fields=('tenant_count',),
-)
-class NestedTenantGroupSerializer(WritableNestedSerializer):
-    tenant_count = serializers.IntegerField(read_only=True)
-    _depth = serializers.IntegerField(source='level', read_only=True)
-
-    class Meta:
-        model = TenantGroup
-        fields = ['id', 'url', 'display_url', 'display', 'name', 'slug', 'tenant_count', '_depth']
-
 
 class NestedTenantSerializer(WritableNestedSerializer):
 
@@ -40,18 +34,6 @@ class NestedTenantSerializer(WritableNestedSerializer):
 #
 # Contacts
 #
-
-@extend_schema_serializer(
-    exclude_fields=('contact_count',),
-)
-class NestedContactGroupSerializer(WritableNestedSerializer):
-    contact_count = serializers.IntegerField(read_only=True)
-    _depth = serializers.IntegerField(source='level', read_only=True)
-
-    class Meta:
-        model = ContactGroup
-        fields = ['id', 'url', 'display_url', 'display', 'name', 'slug', 'contact_count', '_depth']
-
 
 class NestedContactRoleSerializer(WritableNestedSerializer):
