+++ netbox/tenancy/api/serializers.py
@@ -1,3 +1,2 @@
 from .serializers_.tenants import *
 from .serializers_.contacts import *
-from .nested_serializers import *
