+++ netbox/users/api/serializers.py
@@ -1,4 +1,3 @@
 from .serializers_.users import *
 from .serializers_.permissions import *
 from .serializers_.tokens import *
-from .nested_serializers import *
