+++ netbox/virtualization/api/serializers.py
@@ -1,3 +1,2 @@
 from .serializers_.clusters import *
 from .serializers_.virtualmachines import *
-from .nested_serializers import *
