+++ netbox/vpn/api/serializers.py
@@ -1,4 +1,3 @@
 from .serializers_.crypto import *
 from .serializers_.tunnels import *
 from .serializers_.l2vpn import *
-from .nested_serializers import *
