+++ netbox/wireless/api/nested_serializers.py
@@ -1,8 +1,8 @@
-from drf_spectacular.utils import extend_schema_serializer
-from rest_framework import serializers
+import warnings
 
 from netbox.api.serializers import WritableNestedSerializer
 from wireless.models import *
+from .serializers_.nested import NestedWirelessLANGroupSerializer, NestedWirelessLinkSerializer
 
 __all__ = (
     'NestedWirelessLANSerializer',
@@ -10,28 +10,15 @@ __all__ = (
     'NestedWirelessLinkSerializer',
 )
 
-
-@extend_schema_serializer(
-    exclude_fields=('wirelesslan_count',),
+# TODO: Remove in v4.2
+warnings.warn(
+    f"Dedicated nested serializers will be removed in NetBox v4.2. Use Serializer(nested=True) instead.",
+    DeprecationWarning
 )
-class NestedWirelessLANGroupSerializer(WritableNestedSerializer):
-    wirelesslan_count = serializers.IntegerField(read_only=True)
-    _depth = serializers.IntegerField(source='level', read_only=True)
-
-    class Meta:
-        model = WirelessLANGroup
-        fields = ['id', 'url', 'display_url', 'display', 'name', 'slug', 'wirelesslan_count', '_depth']
 
 
 class NestedWirelessLANSerializer(WritableNestedSerializer):
 
     class Meta:
         model = WirelessLAN
-        fields = ['id', 'url', 'display_url', 'display', 'ssid']
-
-
-class NestedWirelessLinkSerializer(WritableNestedSerializer):
-
-    class Meta:
-        model = WirelessLink
         fields = ['id', 'url', 'display_url', 'display', 'ssid']
