+++ netbox/wireless/api/serializers.py
@@ -1,3 +1,2 @@
 from .serializers_.wirelesslans import *
 from .serializers_.wirelesslinks import *
-from .nested_serializers import *
